+++ packages/orq-rc/src/models/operations/retrievedatapoint.ts
@@ -834,7 +834,7 @@ export const RetrieveDatapointResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2025-04-15T21:24:24.287Z",
+    "2025-04-16T07:30:59.029Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {
@@ -877,7 +877,7 @@ export const RetrieveDatapointResponseBody$outboundSchema: z.ZodType<
   createdById: z.string().optional(),
   updatedById: z.string().optional(),
   created: z.date().transform(v => v.toISOString()).optional(),
-  updated: z.date().default(() => new Date("2025-04-15T21:24:24.287Z"))
+  updated: z.date().default(() => new Date("2025-04-16T07:30:59.029Z"))
     .transform(v => v.toISOString()),
 }).transform((v) => {
   return remap$(v, {