+++ packages/orq-rc/RELEASES.md
@@ -1389,3 +1389,13 @@ Based on:
 - [typescript v3.4.0-rc.26] packages/orq-rc
 ### Releases
 - [NPM v3.4.0-rc.26] https://www.npmjs.com/package/@orq-ai/node/v/3.4.0-rc.26 - packages/orq-rc
+
+## 2025-04-16 07:32:26
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.533.0 (2.578.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.4.0-rc.27] packages/orq-rc
+### Releases
+- [NPM v3.4.0-rc.27] https://www.npmjs.com/package/@orq-ai/node/v/3.4.0-rc.27 - packages/orq-rc